flexible-graphrag/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
cmislib
python-alfresco-api==1.1.1
docling
neo4j
numpy
rapidfuzz
spacy
openai
ollama
pydantic
pydantic-settings
# LlamaIndex core and integrations
llama-index
llama-index-llms-openai
llama-index-llms-ollama
llama-index-llms-gemini
llama-index-llms-azure-openai
llama-index-llms-anthropic
llama-index-embeddings-openai
llama-index-embeddings-ollama
llama-index-embeddings-azure-openai
llama-index-vector-stores-neo4jvector
llama-index-vector-stores-qdrant
llama-index-vector-stores-elasticsearch
llama-index-vector-stores-opensearch
llama-index-graph-stores-neo4j
llama-index-graph-stores-kuzu
llama-index-retrievers-bm25
# Additional dependencies
qdrant-client
kuzu
elasticsearch
opensearch-py
